Gate oracle semantics and split semantic facade
